docs(example): Split string into main/parser
